+++ jest.config.js
@@ -1,0 +1,10 @@
+module.exports = {
+  moduleDirectories: ['node_modules', 'src'],
+  moduleFileExtensions: ['js', 'json', 'ts', 'tsx'],
+  rootDir: 'packages/slate-plugins/src',
+  testEnvironment: 'jsdom',
+  testRegex: '(test|spec).tsx?$',
+  transform: {
+    '^.+\\.(t|j)sx?$': 'ts-jest',
+  },
+};