+++ packages/slate-plugins/jest.config.js
@@ -1,0 +1,6 @@
+const jestConfig = require('../../jest.config');
+
+module.exports = {
+  ...jestConfig,
+  rootDir: 'src',
+};