+++ packages/slate-plugins/src/deserializers/index.ts
@@ -1,1 +1,2 @@
 export * from './htmlDeserialize';
+export * from './utils';