+++ packages/slate-plugins/src/deserializers/utils/deserializeBreak.ts
@@ -1,0 +1,3 @@
+export const deserializeBreak = (node: HTMLElement | ChildNode) => {
+  if (node.nodeName === 'BR') return '\n';
+};