+++ packages/slate-plugins/src/deserializers/utils/getDeserializer.spec.ts
@@ -1,0 +1,12 @@
+import { getDeserializer } from './getDeserializer';
+
+describe('when getDeserializer', () => {
+  it('should ', () => {
+    const cb = () => ({ type: 'p' });
+
+    expect(getDeserializer('p', ['p', 'P'], cb)).toEqual({
+      p: cb,
+      P: cb,
+    });
+  });
+});