+++ packages/slate-plugins/src/deserializers/utils/index.ts
@@ -1,0 +1,6 @@
+export * from './getDeserializer';
+export * from './deserializeBreak';
+export * from './deserializeElement';
+export * from './deserializeFragment';
+export * from './deserializeMarks';
+export * from './deserializeTextNode';