+++ packages/slate-plugins/src/elements/blockquote/deserializeBlockquote.ts
@@ -1,10 +1,9 @@
 import { DeserializeHtml } from 'deserializers/types';
+import { getDeserializer } from 'deserializers/utils';
 import { BLOCKQUOTE } from './types';
 
 export const deserializeBlockquote = ({
   typeBlockquote = BLOCKQUOTE,
 } = {}): DeserializeHtml => ({
-  element: {
-    BLOCKQUOTE: () => ({ type: typeBlockquote }),
-  },
+  element: getDeserializer(typeBlockquote, ['BLOCKQUOTE']),
 });