+++ packages/slate-plugins/src/elements/code/deserializeCode.ts
@@ -1,8 +1,7 @@
 import { DeserializeHtml } from 'deserializers/types';
+import { getDeserializer } from 'deserializers/utils';
 import { CODE } from './types';
 
 export const deserializeCode = ({ typeCode = CODE } = {}): DeserializeHtml => ({
-  element: {
-    PRE: () => ({ type: typeCode }),
-  },
+  element: getDeserializer(typeCode, ['PRE']),
 });