+++ packages/slate-plugins/src/elements/heading/deserializeHeading.ts
@@ -1,4 +1,5 @@
 import { DeserializeHtml } from 'deserializers/types';
+import { getDeserializer } from 'deserializers/utils';
 import { DeserializeHeadingOptions, HeadingType } from './types';
 
 export const deserializeHeading = ({
@@ -10,15 +11,18 @@ export const deserializeHeading = ({
   typeH5 = HeadingType.H5,
   typeH6 = HeadingType.H6,
 }: DeserializeHeadingOptions = {}): DeserializeHtml => {
-  const headingTags: any = {
-    H1: () => ({ type: typeH1 }),
-  };
+  let headingTags = getDeserializer(typeH1, ['H1']);
 
-  if (levels >= 2) headingTags.H2 = () => ({ type: typeH2 });
-  if (levels >= 3) headingTags.H3 = () => ({ type: typeH3 });
-  if (levels >= 4) headingTags.H4 = () => ({ type: typeH4 });
-  if (levels >= 5) headingTags.H5 = () => ({ type: typeH5 });
-  if (levels >= 6) headingTags.H6 = () => ({ type: typeH6 });
+  if (levels >= 2)
+    headingTags = { ...headingTags, ...getDeserializer(typeH2, ['H2']) };
+  if (levels >= 3)
+    headingTags = { ...headingTags, ...getDeserializer(typeH3, ['H3']) };
+  if (levels >= 4)
+    headingTags = { ...headingTags, ...getDeserializer(typeH4, ['H4']) };
+  if (levels >= 5)
+    headingTags = { ...headingTags, ...getDeserializer(typeH5, ['H5']) };
+  if (levels >= 6)
+    headingTags = { ...headingTags, ...getDeserializer(typeH6, ['H6']) };
 
   return {
     element: headingTags,