+++ packages/slate-plugins/src/elements/image/deserializeImage.ts
@@ -1,13 +1,12 @@
 import { DeserializeHtml } from 'deserializers/types';
+import { getDeserializer } from 'deserializers/utils';
 import { IMAGE } from './types';
 
 export const deserializeImage = ({
   typeImg = IMAGE,
 } = {}): DeserializeHtml => ({
-  element: {
-    IMG: (el) => ({
-      type: typeImg,
-      url: el.getAttribute('src'),
-    }),
-  },
+  element: getDeserializer(typeImg, ['IMG'], (el) => ({
+    type: typeImg,
+    url: el.getAttribute('src'),
+  })),
 });