+++ packages/slate-plugins/src/elements/paragraph/deserializeParagraph.ts
@@ -1,10 +1,9 @@
 import { DeserializeHtml } from 'deserializers/types';
+import { getDeserializer } from 'deserializers/utils';
 import { PARAGRAPH } from './types';
 
 export const deserializeParagraph = ({
   typeP = PARAGRAPH,
 } = {}): DeserializeHtml => ({
-  element: {
-    P: () => ({ type: typeP }),
-  },
+  element: getDeserializer(typeP, ['P']),
 });